[chiselsim] Add ChiselSettings
Add a new case class, `ChiselSettings`, which will be the main structure
for configuring the Chisel/FIRRTL/Verilog ABI parts of a simulation.
While this may overlap with common compilation settings or backend
compilation settings, this is intended to provide the Chisel user view of
these as opposed to exposign the low-level details of the latter two.

Currently, this only includes layer control.  This is part of a patch
series that will replace existing arguments of layer control type to use
this new Chisel settings structure.  In later patches, new members will be
added to Chisel settings.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>