[chiselsim] Add runUntilFinished stimulus
Add a new object, `stimulus`, that packages up canned Chiselsim stimulus
snippets that can be reused.

Add one stimulus to this, `runUntilFinished`, which will run a simulation
and expect it to terminate naturally (via `$finish` or similar).  If the
simulation runs, but does not finish, then throw a timeout-specific
exception.

This is added as a better way to describe things that the unpublished
ChiselSpec and its methods are doing.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>