[chiselsim] Add runUntilFinished stimulus
Add a new package, `chisel3.simulator.stimulus`, that contains canned
stimulus for use with ChiselSim.

Add one stimulus to this package, `RunUntilFinished`, which will run a
simulation and expect it to terminate naturally (via `chisel3.stop`, i.e.,
a Verilog `$finish`).  If the simulation runs, but does not finish, then
throw a timeout-specific exception.  Use a type class pattern for this, as
suggested by @jackkoenig with implementation generators for `Module`, when
provided with a timeout, and any type when provided with a timeout and a
function to get a `Clock`.

This moves towards removing the unpublished ChiselSpec (which is currently
part of the tests), but forms a core component of how users are expected
to use ChiselSim.  This works towards providing a published replacement
for ChiselSpec that avoids downstream users having to effectively
replicate ChiselSpec.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>