[svsim] End Verilator simulation on $finish
Fix a bug in how Verilator simulations work in svsim where the simulation
would _not_ terminate when a `$finish` was called.  This seems to be a bug
in Verilator where it will not terminate the simulation (as mandated by
the Verilog spec) unless you are running with time delays.

Fixes #4700.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>