[svsim] Add directory control to compilation
Add the ability for a `Workspace` to control which directories are visited
in the build directory when computing the list of source files.  This
differs from the previous behavior which only allowed for file-level
control.  (This continues to preserve file-level control.)

This is infrastructure necessary to prevent compilation of Chisel extract
layer collateral when the extract layer is _not_ included in the
build. See: #4676 for more information.

Co-authored-by: Jack Koenig <koenig@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>