[chiselsim] Exclude disabled extract layers
Change Chiselsim to only include enabled layer directories in compilation.
This changes the previous, problematic behavior where all extract layer
modules were included in compilation and only the ABI file was used to
enable the layer.  The problematic case occurred when the extract layer
included constructs which the target simulator could not
support.  (Specifically, we had problems with Verilator and
`s_eventually`.)

Fixes #4676

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>