Add RequireProperty and EnsureProperty (#4681)
Add support for CIRCT's `circt_verif_require` and `circt_verif_ensure`
intrinsics. These interact with the formal contracts implemented in
CIRCT and translate to asserts or assumes depending on whether the
contract is checked or applied. They lower to asserts if used outside of
a contract.

Chisel itself has no support for contracts yet, such that these two
additions behave like `AssertProperty` in practice. As soon as contracts
are added to Chisel though, they can be used to formulate the pre- and
post-conditions of those contracts.