[svsim] Add arguments to HasSimulator factories
Allow for customizaton of compilation settings used when constructing
simulators via the `HasSimulator.simulators.{verilator, vcs}` type class
factories.  This is added so that general or backend-specific compilation
settings can be passed when the need arises to override the defaults.

This is specifically added to enable the ChiselSpec to ChiselSim where
certain tests will trip Verilator warnings (e.g., `MultiClockSpec` creates
a memory with a multiply driven variable as this is the correct way to
create a synchronizer with a memory).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>