[svsim] Fix VCS run_simulation function (#4726)
Fix an incorrect function signatures for the VCS implementation of
`run_simulation`.  This is a task which has inputs and outputs.  This
needs to be declared as having an `int *` argument for the output in C and
have the input/outputs declared in Verilog.

This should fix an internal error when bumping Chisel w/ VCS-based tests.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>