[test] Fix Vending Machine check
Fix an out-of-bounds vector access that was possible in the vending
machine example.  This would cause the vending machine example to fail
whenever there was a "vend" on the last cycle.  I'm not sure _why_ this
was working before as this bug seems like it was always possible to hit.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>