[test] Add experimental ChiselStage extensions
Add a utility that can be used to replace
`ChiselRunners.elaborateAndGetModule`.  This is intentionally private to
the `chiselTests.experimental` package as I don't want this method to be
used outside of some of the more exotic D/I or DataView APIs which want to
use this.  This also is not something that makes sense (right now) to
expose on `ChiselStage` publicly.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>