[test] Make ChiselFlatSpec non-load bearing
Remove all usages of ChiselFlatSpec.  Replace these with direct usage of
Scalatest.  This is done in order to delete ChiselFlatSpec.

Usages of traits that package up unknown other traits, one-off methods,
etc. is generally bad.  This makes it so that individual tests are opaque
and require an understanding of the inheritance hierarchy.  This makes
each test self-contained.  It's just Scalatest.

Note: part of this is to defensively prevent methods getting added to
`ChiselFlatSpec` or a trait it mixes in when instead those should be
public methods that any user of Chisel has access to.

I also sneaked in some import reorganization with this commit...

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>