[test] Refactor, fix PortSpec
Refactor PortSpec to use FileCheck and flat spec.  This was the only
remaining test that used `ChiselFreeSpec` which is slated to be deleted.

Fix a possible bug in the test where this was trying to check that the
automatic clock and reset ports do not have source locators.  However,
these _do_ get source locators and the original test was not actually
checking this.  Change the check to lock in the current behavior of these
get source locators.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>