[testing] Add FileCheck API (#4749)
Add an extension-method based FileCheck API.  This builds off of
@jackkoenig's original FileCheck trait inside ChiselSpec and repackages it
for publishing.

This makes a number of deviations from the original version:

  1. This opts for extension methods.
  2. This does not add helpers for running Chisel.
  3. This throws dedicated assertions on failures.

Taking (1) and (2) together, this is intended to _not_ duplicate methods
which are already availble via ChiselStage.  Specifically, I am intending
the API to be used like:

    ChiselStage.emitSystemVerilgo(new Foo).fileCheck()("CHECK: module Foo")

Things are done this way to try to make the APIs as orthogonal as possible
without too many redundant ways to accomplish the same thing.

(3): is added for convenience of testing.  On failure, this provides an
output that approximates what you get from lit (LLVM's Integrated Tester).
This additionally provides a rerun script that you can just copy-paste
into a terminal to reproduce the failure.

For ease of reproduction, both the input and the check are written to
files with predictable names in a temporary directory underneath what is
provided by an implementation of the `HasTestingDirectory` type class.
Having both the input and check file are necessary to do easy reruns.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>