[simulator] Remove DefaultSimulator
Remove `DefaultSimulator`.  This is fully replaced with the `ChiselSim`
API and is no longer needed.  Notably, `DefaultSimulator` never made it
onto an actual release of Chisel so this is safe to just remove.

Rewrite EphemeralSimulator in terms of ChiselSim APIs.  Migrate tests of
DefaultSimulator to tests on ChiselSim or other simulator classes,
objects, and methods.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>