[svsim] Centralize simulation-driver global state
Change simulation-driver.cpp to centralize its global, mutable state into
a single struct, `State`, and a single static pointer that is initizlied
by the `main` function.  This isn't awesome, but it at least avoids the
scattering of  global state that existed before.

Actually removing the global state is hard because this global state needs
to be accessible by functions that are exported via DPI and called from
Verilog, namely, `simulation_body`.  Or: there is an eventual refactor
that avoids the global state, but it may be more invasive and I'd like to
do that at a later time.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>