[svsim] Add clangformat target for svsim (#4765)
Add a mill target to the svsim project which can be used to run a basic
clang-format of the C++ resource file.  I'm adding this as I'm planning to
do some heavier work on this file and don't want to be worrying about
formatting.

This is not added to CI as this is just one file and can be run manually
for now.

Co-authored-by: Jack Koenig <koenig@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>