Add formal contracts (#4682)
Add the `Contract` API which allows users to express the pre and
post-conditions of a piece of hardware. These can be used during formal
verification to prove that the hardware upholds the conditions laid out
in the contract. Additionally, contracts may be used to implement a
divide-and-conquer strategy to simplify larger proofs since they act
like a cutpoint and an assume-guarantee substitution.