[svsim] Avoid EndOfMessages Exception
Add a `final` block to svsim's Verilog which will call a DPI function to
finish the simulation.  This causes the simulator to either exit (if a
done command has been received) or wait until it sees _excatly_ a done
command.

Because of the different ways that simulators can choose to implement
time delays, the simulator _may_ have outstanding transactions that need
to be finished.  Finish any outstanding transactions before blocking on a
done command.  Use this new outstanding transaction state to implement the
old `receivedDone` state.

This is done to get equivalent behavior between Verilator and VCS when a
Verilog `$finish` is executed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>