Change scalatest.WithTestingDirectory directory
Change the directory that Scalatest tests use from being rooted at
`build/` to being rooted at `build/chiselsim/`.  The former can have a lot
of contention with many output files and it would be better to add this
additional level.  Also, this aligns this with the new default behavior of
the non-Scalatest WithTestingDirectory.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>