[scalatest] Add subdirectory control
Add a new overridable member function to `scalatest.TestingDirectory`
which allows a user to set a subdirectory for a test.  The primary use
case for this is to allow the user to set a directory derived from a seed
for a test (or a timestamp).  This is not the standard behavior.
Normally, you only have one output directory and you are fine to override
it.  However, we have internal uses where these directories need to be
split up based on seed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>