[testing] Make TestingDirectory stackable
Fix a bug in the `TestingDirectory` trait where this was not "stackable"
with other traits.  This needs to use an exact pattern shown on the
Scalatest documentation where the `withFixture` of the superclass is
called so that all fixtures will be applied.  See "Composing fixtures by
stakcing traits" in Scalatest docs [[1]].

[1]: https://www.scalatest.org/user_guide/sharing_fixtures#withFixtureNoArgTest

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>