[testing] Add HasConfigMap Scalatest trait
Add a new, stackable trait called `HasConfigMap` which exposes Scalatest's
`configMap` [[1]] in an easy-to-use way.  This allows for users to pass
arguments on the command line to Scalatest via the `-D<key>=<foo>` option
and have access to them in a `Map[String, Any]` in their test.

While I am generally not hugely in favor of this pattern, as it can be
abused to make tests pass or fail via options when I would prefer that
tests were self-contained, there are some internal usages of this for
doing things like changing the simulator, turning on waveforms, or
extending the length of randomized testing.  Given that, it's a useful
feature to have and I would prefer if users went rolling their own
solutions to this.

[1]: https://www.scalatest.org/scaladoc/3.1.0/org/scalatest/ConfigMap.html

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>