[testing] Add HasTestingDirectory subdir factory (#4774)
Add a new member function to `HasTestingDirectory` called
`withSubdirectory`.  This is a factory of _new_ `HasTestingDirectory`
objects which will return directories nested under the original.

This is intended to fit a very specific use case where a user is passing
command line arguments through Scalatest's ConfigMap API (i.e.,
`-D<key>=<value>`) _and they want to use the information in the ConfigMap
to affect the directory structure_.  Because the ConfigMap is only
availalbe _inside the test_, it cannot be used to set the test name.  If a
user is not passing arguments via the ConfigMap, then a user can instead
structure the ScalaTest scopes in such a way to get the directory
structure they want.

Co-authored-by: Jack Koenig <koenig@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>