[chiselsim] Add CLI option support
Add Scalatest integration for Chiselsim which allows users to write traits
which add command line options to a Scalatest test.  These hook into the
`-D` Scalatest options (i.e., the ConfigMap) and allow the user to modify
the common or backend-specific options of a test.  This includes
validation features to guarantee that illegal options are not passed.

Add an example trait, `WaveformVCD` which will enable VCD dumping for
_either_ Verilator or VCS.  With this, you can then do, for a test which
adds this trait:

    ./mill 'chisel3[2.13.16].test.testOnly' 'FooSpec' -- -DemitVCD=true

This is related to #4770.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>