[svsim] Use SNPS-specific license in VCS
Fix the logic for grabbing environment variables for the replay Makefile
when running VCS tests.  The VCS manual says that you grab the
Synopsys-specific license file (SNPSLMD_LICENSE_FILE) first and only use
the generic license file (LM_LICENSE_FILE) if that fails.  The existing
code was wrong in that it only grabbed the latter.  There is no
requirement that the latter is available and that is only a fallback if
the former doesn't exist.

Also, add some lightweight infrastructure to cleanup the representation of
VCS license files.

I tested this internally and all the tests passed.  (I had to change the
test to queue for licenses, though...)

Fixes #4785.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>