[chiselsim] Chisel/firtool Options on SimulatorAPI (#4794)
Modify ChiselSim simulate methods to allow for Chsiel and firtool options
to be passed to the `simulate` methods.  This is done to allow for things
like `--throw-on-first-error` to be passed.

For more context, we have a number of CLI options internally for working
around the fact that ChiselSim doesn't support passing the Chisel/firtool
compilation options to ChiselStage.  This seems wrong to me as I want to
have the SimulatorAPI methods be richer externsions of what you can do
with ChiselStage.  Ergo, because ChiselStage can control Chisel and
firtool options, so should the SimulatorAPI.

The only slightly sharp edge here is that certain Chisel options may be
illegal.  The Simulator APIs are already going to compile your module to
SystemVerilog.  If you were to try to use `--target chirrtl`, then that is
an illegal option combination.  This seems fine with the understanding
that this can happen and users need to know what they are doing.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>