[svsim] Add plusarg support
Add the ability to pass plusargs to svsim simulations.

Note: there is some ugliness here in that simulation options are passed to
functions that are doing "compilation".  This is a software architecture
problem in svsim as evidenced by backend compilation options including
runtime options (for VCS).  This all needs to be cleaned up separately.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>