[chiselsim] Fix Settings Mods. unapplied bug
Fix a bug where modifications applied via command line arguments would not
be reflected by ChiselSim.  This was an issue of not using the modified
settings.

This additionally changes it so that the CLI modifications are always
applied _last_.  I.e., they could overwrite anything earlier or specified
in Chisel.  This seems to make the most sens as I would like to, as a
user, be able to change the behavior of a test on the command line.
Nothing else should override this.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>