[chiselsim] mv HasCliSimulator into Cli.Simulator
Change the `HasCliSimulator` trait to instead be implemented using
`HasCliOptions` infrastructure.

This fixes a bug where the old `HasCliSimulator` could not be used
together with `HasCliOptions` because the latter enforces that only legal
options exist.  `HasCliSimulator` was a way to add options that were not
registered with `HasCliOptions` making them non-composable.

This has the benefit of the simulator option will now show up along with
the standard `HasCliOptions` help text.

The above "bug" can happen for any test that is trying to add options that
are not registered.  This is more of a feature than a bug as it forces
options for a test to be discoverable while guaranteeing that we can check
for illegal options.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>