[testing] Fix testing dir date/time stability
Fix a bug in the timestamp and temporary implementations of the
`HasTestingDirectory` type class.  These need to, for a given type class
implementation, always return the same date/time used when constructing a
temporary directory.

This fixes a bug if a user is trying to get the current directory, e.g.,
via an `implicitly[HasTestingDirectory]` call.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>