[chiselsim] Add waves at time zero to settings
Add an option to ChiselSim settings to enable waves at time zero.  While
this is now the _fourth_ way that this can be done (enableWaves,
svsim.CommonCompilationSettings, and via modifications type classes to
CommonCompilationSettings/via the CLI), this is useful/necessary to have a
place that waves can be turned on from inside a test that still works with
the `simulate` API and its reset procedure.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>