[svsim] Add PlusSeparated.any
Add a member function to the `PlusSeparated` trait called `any` which will
return true if any of the elements are set.  This is useful for writing
Cli handlers which want to know if any value is set and then set some
other option.  (This came up in an internal usage where if any VCS
coverage option is set, then certain preprocessor defines are also set.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>