[svsim] Add VCS flags member, pass to invocation
Add a sequence of `Flag.Type` which can be given to VCS.  This is intended
as a catch-all for all no-argument flags which should be passed to VCS
directly.  This is done this way to avoid blowing up the VCS compilation
settings with one field-per-option.  This way, more `Flag.Type` can just
be added and used.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>