[chiselsim] Add Cli.Scale trait
Add a new trait `Cli.Scale` which can be used for making tests that should
be command-line scalable.  This adds a `-Dscale=<double>` option and
provides a member function `scaled(Int)` which can be used to scale some
property of a test, like its runtime.

This is a replacement for an internal flow which re-used the `-F` option
that Scalatest provides to mean something different from Scalatest's
original intent.  This is changed here so that the existince of a scaling
factor is documented as part of the help text of a test.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>