[chiselsim] Require help text in Scale trait
Change the `Scale` trait (which adds the `-Dscale=<float>` Scalatest
parameter) to require the user to provide documentation about what the
scale is for that test.

This is done to avoid potential confusion about what the "scale" means for
a given test.  Does it lengthen the test?  Does it test more aggressively?
Does it instantiate more more units?  Does it change latencies?

While it is entirely reasonable for the "scaling factor" to mean different
things for different tests, I want test authors to be forced to say what
it means for their test and to then make this both discoverable in the
code and discoverable on the command line (via `-Dhelp=1`).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>