[chiselsim] Add RunUntilSuccess stimulus
Add new stimulus, `RunUntilSuccess`, which can be used to run a simulation
until a user-specified "success" port asserts.  This is added as a
complementary testing approach to `RunUntilFinished` whish is needed for
some environments where the test harness cannot terminate on its own or
the test harness is re-used in a separate, non-ChiselSim testing context
where it cannot terminate on its own.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>