[chiselsim] Add randomization to settings
Add the ability to control the existing FIRRTL randomization settings via
an additional parameter to `chisel3.simulator.Settings`.  Set the default
behavior for this to always randomize.  This is a reasonable default for
Verilator or any other 2-state simulator.  For 4-state simulators, users
may want to deviate and use the provided `uniniitialized` randomization.

This commit intentionally does not try to struggle with setting
per-simulator settings that are different from each other.  While I think
this could make sense (you would like to use randomization for Verilator
and uninitialized w/ x-prop for VCS), I think that the logic necessary to
handling this is going to be too opaque for most users.  I would rather
keep this as a single setting on simulate and any simulator you use will
then simulate the same thing.  If you want to do something different, then
it is on the user.

Note: I don't expect most users to care about running different
simulators, in spite of the fact that we provide the `Cli.Simulator`
trait (and intentionally do not mix it into `ChiselSim`).  If users _do_
care about running different simulators, then they would be better served
by enumerating these as explicit tests.  I.e., I am not super concerned
about the use case of writing a test with one randomization mode and then
expecting the randomization mode to morph if the user changes the
simulator on the command line.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>