Bump emitted FIRRTL to 5.1.0
Bump the emitted FIRRTL version to 5.1.0.  This is done because we want to
start using the new `{{SimulationTime}}` printf construct.  This is a
major change to FIRRTL (as it is a backwards incompatible change) and
therefore we have to go through the major bump.

The minor bump needs to stay at "1" because we are still emitting things
which are not in the FIRRTL spec, e.g., Path ops.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>