[svsim] Use -no_save instead of ASLR workarounds (#4845)
For the internal installation of VCS and RHEL, our simulations do not work
due to VCS' save/restore functionality which will launch a simulation,
detect if ASLR is present, and then kill the simulation and relaunch it
with ASLR disabled.  The workaround employed here has been empirically
shown to be problematic for at least one user (#4837).  Drop the ASLR
workaround and add the runtime option `-no_save` which turns off VCS
save/restore functionality.  svsim doesn't support this and it's ill
advised to use non-standard workarounds.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>