Add SimulationTest marker
The FIRRTL spec is about to define a `simulation` construct to indicate
that a module should be executed as a simulation test. Add a user-facing
`SimulationTest` class that can be used to mark a module as to be
executed as a simulation test. This would commonly be used inside a test
harness module to mark the surrounding module as a test. For example:

    class TestHarness extends RawModule {
      SimulationTest(this)
    }

The simulation test can be given an optional name and a map of
parameters. The parameters are identical to the ones accepted by the
`FormalTest` marker. Multiple simulation test markers may be added to a
single module, which may be useful if a test should be run with
different sets of user-defined parameters.