[FIRRTL] Serialize Named as Target (#4853)
Change the way that `Named` is serialized into JSON to have this serialize
as a `Target`.  This prevents a legacy serialization from ever showing up
without actually removing `Named`.

This is done to free up the ability for us to change the serialization of
`Target` to drop the circuit name without introducing a collision with the
`Named` syntax.

E.g., I want to change the syntax for a target from:

    ~Foo|Foo

To:

    Foo

However, this aliases with a `CircuitName` of `Foo`.  However, by ensuring
that everything is funneled through `Target`, this can then be changed on
a FIRRTL version bumpd without worrying about backwards compatibiliy with
`Named`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>