Deprecate Printable.unpack and .unpackArgs, replace with new version (#4858)
The old version conflated the concerns of lowering to FIRRTL format
Strings with what should simply be the inverse operation of
Printable.pack. Specifically, Name and FullName are not represented in
FIRRTL so must be "erased" before lowering to FIRRTL. This removal
requires use of internal Chisel types (Component) so was always weird to
have on a public API. Rather than doing this in .unpack, we instead
treat that as a separate operation.