[firrtl] Remove Named.serialize
Remove the ability to serialize a `Named`.  After 9532db555, all `Named`
are converted to targets before being serialized.  This flat out avoids
the problem of any `Named` ever being serialized using its legacy syntax
by just not defining a serialization.

This is done as part of removing the circuit from targets to enable inline
unit testing to deal with `Definition`s which are created outside of a
circuit context.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>