[firrtl] Remove `Named.serialize` and miscellaneous APIs (#4861)
Remove the ability to serialize a `Named`.  After 9532db555, all `Named`
are converted to targets before being serialized.  This flat out avoids
the problem of any `Named` ever being serialized using its legacy syntax
by just not defining a serialization.

This is done as part of removing the circuit from targets to enable inline
unit testing to deal with `Definition`s which are created outside of a
circuit context.

Drop an unused API for creating RenameMaps.  This was used by the SFC, but
is now completely private and has no users.

Remove two Annotation-related utilities which have no in-tree users.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>