Add --split-verilog to emitSystemVerilogFile (#4867)
Change the behavior of `ChiselStage.emitSystemVerilogFile` to add
`--split-verilog`.  Without this, a user can run into a situation
where (rightly), they get non-Verilog output in the single file output.
Single file output from CIRCT is only intended to be used for testing.

Instead of telling users that they have to add "--split-verilog", just add
it by default with no way to turn it off.  I don't see a reason why a user
_should_ be able to turn this off as the "off" mode is only for testing
and this testing can be just as easily done with `emitSystemVerilog`.

Fixes #3933.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>