[chiselsim] Add a "flag" CLI option factory (#4880)
Add another factory for creating options that is intended to be used for
creating "flag"-style options.  This exists to unify this pattern as
opposed to forcing users to invent their own when they need it.

As can be seen, this unifies a number of existing flag-like options
relating to waveform emission.

CC: @andrew-gouldey-sifive: This will be available in the next internal
Chisel release for your use.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>