[testing] Mangle more directories (#4927)
Directories created with `[;<>]` are problematic for Makefiles and/or
Verilator.  Drop these if we see them.  This fixes issues where tests,
converted to ChiselSim, will fail to compile through Verilator due to
observed use of semicolons in test scopes or serialization of Chisel
types, e.g., `UInt<16>`.

Note: the continued patching of this indicates that we should setup an
allowlist of legal characters as opposed to continuing to exlcude
characters.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>