Revert "[chiselsim] Add Randomization.x (#4939)"
This reverts commit 2cb90ea29e43bf2805b5a084e8fd4f734a47b5c4.

After thinking about this commit more, there is no need to add an explicit
x randomization.  This is the same as the "uninitialized" randomization.
If this had an effect in Verilator, or there was some utility in doing a
non-random simulator randomization followed by an x initialization, then
this might make sense.  However, at this time, this doesn't do anything.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>