[chiselsim] Add frequency control to stimulus
Add the ability to better control the frequency of a simulation in
ChiselSim stimulus.  Previously, this was left up to the simulator to
define and could result in some weird behaviors: Verilator defaults to
`1ps/1ps` (which would cause simulations to seemingly simulate very fast)
and VCS defaults to `1s/1s` (which would cause simulations to seemingly
simulate very slowly).

Change the default behavior to simulate at 1GHz.

Note: this is being done because some gate-level sims are being done using
ChiselSim and accurate power requires realistic frequencies. If users
don't care about this (or aren't integrating "real" models of hardware)
then this change shouldn't matter or affect users other than seeing the
units in the favorite waveform viewer change.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>