[svsim] Set "+all" for VCS FSDB dump
Change the behavior of VCS FSDB dumping to dump _all_ signals.  This
includes things like multi-dimensional arrays which are useful for power
simulations.

If this becomes problematic in the future due to size, we can figure out a
way to provide more fine-grained control of what is dumped.  This would
then likely be better handled with a VCS simulation setting.  (VCS/Verdi
provide multiple different ways that you can turn this on: environment
variable, command line, or via a system function.).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>