[chiselsim] Stop using CIRCT filelists
Change ChiselSim to stop using undocumented CIRCT-generated filelists to
know what files to include in simulator compile.  Specifically, don't rely
on the existence of `filelist.f` or `firrtl_black_box_resource_files.f`.

This is needed because the latter file was removed from (the unreleased)
CIRCT 1.125.0 [[1]] for reasons that it wasn't supposed to be load
bearing.  Removing the reliance on the former is just good to do.

Note: this may result in some breakages as users may have been
inadvertently relying on this behavior to exclude some generated files
from their compilation.  E.g., a user could generate some Chisel-time
metadata file which gets put into the build area, but then is excluded
from the compile.  To work around this, users should dump these files in
an alternative area.

[1]: https://github.com/llvm/circt/commit/c7e10d7c35f860c406a342ef54d21c84a132c7a3

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>