[chiselsim] Use allowlist for compilation includes
Change the logic used for determining which files to incldue from a
blocklist to an allowlist.  The blocklist approach is problematic as
sometimes users will generate odd files (e.g., '.graphml') and these need
to be excluded from the build.  It is easier to just include reasonable
files which is all Verilog and C++ files.

This is an imperfect solution and should be changed in the future to
either: (1) allow customization of this, (2) split the Chisel build
directory from the FIRRTL build directory, or (3) use CIRCT tooling to
query the MLIR to get the full list of files that should be included.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>